Add app config and ORM mapping.
